+++ samples/Sentry.Samples.Android/MainActivity.cs
@@ -1,3 +1,5 @@
+using Sentry.Android;
+
 namespace Sentry.Samples.Android;
 
 [Activity(Label = "@string/app_name", MainLauncher = true)]
@@ -9,6 +11,8 @@ public class MainActivity : Activity
         {
             o.Dsn = "https://eb18e953812b41c3aeb042e666fd3b5c@o447951.ingest.sentry.io/5428537";
             o.SendDefaultPii = true; // adds the user's IP address automatically
+            o.Android.LogCatIntegration = LogCatIntegrationType.Errors; // Get logcat logs for both handled and unhandled errors; default is unhandled only
+            o.Android.LogCatMaxLines = 1000; // Defaults to 1000
         });
 
         // Here's an example of adding custom scope information.