+++ src/Sentry/Platforms/Android/SentryOptions.cs
@@ -1,3 +1,5 @@
+using Sentry.Android;
+
 // ReSharper disable once CheckNamespace
 namespace Sentry;
 
@@ -195,6 +197,20 @@ public partial class SentryOptions
         /// The default value is 5 seconds.
         /// </summary>
         public TimeSpan ReadTimeout { get; set; } = TimeSpan.FromSeconds(5);
+
+        /// <summary>
+        /// Gets or sets whether when LogCat logs are attached to events.
+        /// The default is <see cref="LogCatIntegrationType.None"/>
+        /// </summary>
+        /// <seealso cref="LogCatMaxLines" />
+        public LogCatIntegrationType LogCatIntegration { get; set; } = LogCatIntegrationType.None;
+
+        /// <summary>
+        /// Gets or sets the maximum number of lines to read from LogCat logs.
+        /// The default value is 1000.
+        /// </summary>
+        /// <seealso cref="LogCatIntegration" />
+        public int LogCatMaxLines { get; set; } = 1000;
 
 
         // ---------- Other ----------